Trapezoidal Drive at TIM2 interrupt
Had to make all the functions static due to 2 reasons:
1- Main one was that I needed to call the Driving function at a larger
frequency (larger than 1kHz) so I needed to call the function in an
interrupt, so I configured TIM2 interrupt for this, but I needed to call
the function from the class, so I needed it to be static.
2- There is no way that we can drive more than one PMSM with this board,
so there is no need to create instances of the driver class, so all the
functions can be static.